requirements.txt
torch
music21
pyyaml